requirements.txt
django
djangorestframework
markdown
django-filter
setuptools_rust
django-rest-auth
django-allauth
openpyxl
